extracted <GameConfigDialog/>
